Create output
Reporting the output of the java code that was implemented regarding the VRP.
The examined Vehicle Routing Problem implemented with the following Greedy Computational Strategy method. The proposed solution to this problem is the sequence of customers within routes. The constraints are the capacity of the vehicles and customers being serve exactly once.